+++ wisp_api/index.ts
@@ -1,101 +1,44 @@
-export type RequestTypes = "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
+import { WispAPICore } from "./apis/index.js";
+import { AllocationsAPI } from "./apis/allocations.js";
+import { AuditLogsAPI } from "./apis/audit_log.js";
+import { BackupsAPI } from "./apis/backups.js";
+import { DatabasesAPI } from "./apis/databases.js";
+import { FastDLAPI } from "./apis/fastdl.js";
+import { FilesystemAPI } from "./apis/filesystem.js";
+import { ModsAPI } from "./apis/mods.js";
+import { SchedulesAPI } from "./apis/schedules.js";
+import { ServersAPI } from "./apis/servers.js";
+import { StartupAPI } from "./apis/startup.js";
+import { SubusersAPI } from "./apis/subusers.js";
 
-export type PaginationData = {
-  total: number;
-  count: number;
-  perPage: number;
-  currentPage: number;
-  totalPages: number;
-}
+export class WispAPI {
+  private core: WispAPICore;
 
-export interface WispAPICore {
-  domain: string;
-  uuid: string;
-  token: string;
-  logger: any;
-}
+  public Allocations: AllocationsAPI;
+  public AuditLogs: AuditLogsAPI;
+  public Backups: BackupsAPI;
+  public Databases: DatabasesAPI;
+  public FastDL: FastDLAPI;
+  public Filesystem: FilesystemAPI;
+  public Mods: ModsAPI;
+  public Schedules: SchedulesAPI;
+  public Servers: ServersAPI;
+  public Startup: StartupAPI;
+  public Subusers: SubusersAPI;
 
-export class WispAPICore {
   constructor(domain: string, uuid: string, token: string, logger: any) {
-    this.domain = domain;
-    this.uuid = uuid;
-    this.token = token;
-    this.logger = logger;
-  }
+    this.core = new WispAPICore(domain, uuid, token, logger);
 
-  setUUID(uuid: string) {
-    this.uuid = uuid;
-  }
-
-  makeURL(path: string) {
-    return `https://${this.domain}/api/client/servers/${this.uuid}/${path}`;
-  }
-
-  async makeRequest(method: RequestTypes, path: string, data?: any) {
-    let url = this.makeURL(path);
-    const headers = new Headers({
-      "Content-Type": "application/json",
-      "Accept": "application/vnd.wisp.v1+json",
-      "Authorization": `Bearer ${this.token}`,
-      "User-Agent": "WispJS (https://github.com/CFC-Servers/wispjs, 1.0.0)"
-    });
-
-    const request = async () => {
-      let response: Response;
-
-      console.log(`${method} -> ${url}`);
-
-      switch(method) {
-        case "GET":
-          if (data !== null) {
-            const params = new URLSearchParams(data);
-            const uri = new URL(url);
-
-            uri.search = params.toString();
-            url = uri.toString();
-            console.log(`Updated GET URL: ${url}`);
-          }
-
-          response = await fetch(url, { method: "GET", headers: headers });
-          break;
-
-        case "POST":
-          data = JSON.stringify(data);
-          response = await fetch(url, { method: "POST", headers: headers, body: data });
-          break;
-
-        case "PUT":
-          data = data ? JSON.stringify(data) : "";
-          response = await fetch(url, { method: "PUT", headers: headers, body: data });
-          break;
-
-        case "PATCH":
-          data = JSON.stringify(data);
-          response = await fetch(url, { method: "PATCH", headers: headers, body: data });
-          break;
-
-        case "DELETE":
-          response = await fetch(url, { method: "DELETE", headers: headers });
-          break;
-
-        default:
-          throw new Error(`Invalid method: ${method}`);
-      }
-
-      if (!response.ok) {
-        const status = response.status;
-        const statusText = response.statusText;
-        this.logger.error(`Request failed: ${method} -> ${url}: ${status} - ${statusText}`);
-
-        const text = await response.text();
-        this.logger.error(text);
-
-        throw new Error(`Request failed! Status: ${status} - ${statusText}`);
-      }
-
-      return response;
-    }
-
-    return await request();
+    this.Allocations = new AllocationsAPI(this.core);
+    this.AuditLogs = new AuditLogsAPI(this.core);
+    this.Backups = new BackupsAPI(this.core);
+    this.Databases = new DatabasesAPI(this.core);
+    this.FastDL = new FastDLAPI(this.core);
+    this.Filesystem = new FilesystemAPI(this.core);
+    this.Mods = new ModsAPI(this.core);
+    this.Schedules = new SchedulesAPI(this.core);
+    this.Servers = new ServersAPI(this.core);
+    this.Startup = new StartupAPI(this.core);
+    this.Subusers = new SubusersAPI(this.core);
   }
 }
